+++ src/constants/index.ts
@@ -1,3 +1,3 @@
-export * from './storyMetaData';
-export * from './storyStyles';
-export * from './storyDummyData';
+export * as STORY_META from './STORY_META';
+export * as STORY_STYLE from './STORY_STYLE';
+export * as DUMMY from './DUMMY';
