+++ src/hooks/public/useValidationStore.tsx
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 
-import { createValidationStore } from '../utils';
+import { createValidationStore } from '../../utils';
 
 /**
  * @deprecated validationObserver를 사용해주세요.
