--- src/utils/index.tsx
@@ -1,5 +1,0 @@
-export * from './cleanClassName';
-export * from './regex';
-export * from './createValidationStore';
-export * from './dummy';
-export * from './getLabelSpace';
