+++ src/utils/public/cleanClassName.ts
