+++ tsconfig.json
@@ -25,7 +25,7 @@
       "@components/*": ["components/*"],
       "@utils": ["utils"],
       "@hooks": ["hooks"],
-      "@contexts": ["contexts"],
+      "@contexts/*": ["contexts/*"],
       "@constants": ["constants"],
       "@styles/*": ["styles/*"]
     }
